Add files via upload
You have to first disable the google cookies and if one page uploaded is going to be deleted , you can change the number of pages going to be deleted as well. u can set a list of 
countries that u want to explore (because i explored the job from all the countries available in indeed)
the work of code as follows:
>> First, get the the indeed page 
>> Click the Pays link in the footer of the page.
>> after the list of countries shown
>> then the code get the link from each element "li" and presses the pay link to get the job page for the specified country and when all pages of this country exracted ,
The driver gets back to the previous page "driver. get()" line within a loop, and click on the second link, etc.